pyproject.toml
[tool.poetry]
name = "lance"
version = "0.5.0"
description = "LLM Red Team Evaluation Framework — lance.iosec.in"
authors = ["Shekhar Suman <shekhar@iosec.in>"]
readme = "README.md"
license = "MIT"
homepage = "https://lance.iosec.in"
repository = "https://github.com/iosec-in/lance"
keywords = ["llm", "red-team", "security", "ai-security", "prompt-injection"]

[tool.poetry.dependencies]
python = "^3.11"
litellm = "^1.35.0"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
sqlalchemy = "^2.0.0"
alembic = "^1.13.0"
pydantic = "^2.7.0"
pydantic-settings = "^2.2.0"
typer = {extras = ["all"], version = "^0.12.0"}
rich = "^13.7.0"
pyyaml = "^6.0.1"
httpx = "^0.27.0"
jinja2 = "^3.1.4"
weasyprint = "^62.3"
sentence-transformers = "^3.0.0"
presidio-analyzer = "^2.2.354"
presidio-anonymizer = "^2.2.354"
aiofiles = "^23.2.1"
python-multipart = "^0.0.9"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
pytest-asyncio = "^0.23.6"
ruff = "^0.4.4"
mypy = "^1.10.0"

[tool.poetry.scripts]
lance = "lance.cli.main:app"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
